\name{healthvis.plot}
\alias{healthvis.plot}
\title{A method for plotting a healthvis object}
\usage{
  \method{healthvis}{plot} (x,
    file.name = "survivalPlot.html")
}
\arguments{
  \item{healthvisObj}{A healthvis object to be plotted}

  \item{file.title}{The file name for the HTML page to be
  created}
}
\value{
  \code{NULL}
}
\description{
  A method for plotting a healthvis object
}

